move explorercontrolsection to own section
